pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "salai-spk-diar"
version = "0.1.0"
description = "Speaker diarization research for ATC communications — SaLAI Lab, ERAU"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "Shital Pandey", email = "pandeys2@my.erau.edu" },
]

dependencies = [
    # --- Core scientific stack ---
    "numpy>=1.26,<2.0",
    "scipy>=1.12",
    "soundfile>=0.12",
    "librosa>=0.10",

    # --- PyTorch (CUDA 12.4) ---
    "torch>=2.4",
    "torchaudio>=2.4",
    "lightning>=2.3",

    # --- NeMo toolkit (speaker diarization) ---
    "nemo-toolkit[asr]>=2.0",

    # --- Pyannote (clustering-based baseline) ---
    "pyannote.audio>=3.3",

    # --- Evaluation ---
    "pyannote.metrics>=3.2",

    # --- Visualization ---
    "matplotlib>=3.8",
    "seaborn>=0.13",

    # --- Data handling ---
    "pandas>=2.1",
    "datasets>=2.18",

    # --- Config and CLI ---
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "rich>=13.0",

    # --- Utilities ---
    "tqdm>=4.66",
    "optuna>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
    "ipython>=8.0",
    "jupyter>=1.0",
]

[tool.uv]
# PyTorch CUDA 12.4 builds come from their own index
extra-index-url = ["https://download.pytorch.org/whl/cu124"]

[tool.uv.sources]
# Ensure torch/torchaudio resolve from the CUDA index
torch = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
